refactor: rename SkillEvaluation to EvaluationDetails
Generalize evaluation UI components to accept base EvaluationResult type.
Add assertionDescriptions prop for consumer-specific assertion labels.
Rename SCSS module and update all imports.
