+++ src/experiences/SkillEvaluation/index.ts
@@ -1,5 +1,11 @@
-export { EvalScoreBadge } from './EvalScoreBadge';
-export { SkillEvaluationDetails } from './SkillEvaluationDetails';
-export { EvalScoreBar } from './EvalScoreBar';
-export { EvalAssertionList } from './EvalAssertionList';
-export { EvalRadarChart } from './EvalRadarChart';
+/**
+ * @deprecated Import from '@/experiences/EvaluationDetails' instead.
+ * Temporary re-export shim — will be deleted after all imports are migrated.
+ */
+export {
+  EvalScoreBadge,
+  EvaluationDetails as SkillEvaluationDetails,
+  EvalScoreBar,
+  EvalAssertionList,
+  EvalRadarChart,
+} from '../EvaluationDetails';
